third_party/WebKit/LayoutTests/fast/css/invalidation/explicit-inheritance-non-inherited-hover.html

<!DOCTYPE html>
<script src="../../../resources/js-test.js"></script>
<style>
html { background-color: green }
* { background-color: inherit }
.a:hover.b { background-color: green }
</style>
<div id="root">
    <div>
        <span></span>
        <span></span>
        <span></span>
        <span></span>
    </div>
</div>
<script>
description("Inheritance-propagation should stop when there is no style diff and the affected-by-hover flag changes.");

root.offsetTop; // force recalc

root.className = "a b";

if (window.internals)
    shouldBe("internals.updateStyleAndReturnAffectedElementCount()", "1");
</script>
