third_party/WebKit/LayoutTests/fast/css/invalidation/explicit-inheritance-non-inherited-hover-expected.txt

Inheritance-propagation should stop when there is no style diff and the affected-by-hover flag changes.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.updateStyleAndReturnAffectedElementCount() is 1
PASS successfullyParsed is true

TEST COMPLETE
